+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ﻿# Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added shipyards that are building ships to the resource consumption display.
+
 ## [0.1.0] - 2024-03-27
 
 ### Added
 
 - Initial release. Show resource and production statistics in the island storage screen.
 
+[Unreleased]: https://github.com/kepons/pp2-production-stats/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/kepons/pp2-production-stats/releases/tag/v0.1.0